### 2025-04-01 #7972 @warrickball
Hi @KostasValeckas, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. I'll indicate up front that this will probably take several months given the current backlog. Sorry, and thanks for your patience!

In the meantime, I noticed a few things we can tidy up in the paper.

* Please add countries to the affiliations.
* You can group parenthesised citations together by writing e.g. `[@cr_1; @cr_2]` instead of `[@cr_1] [@cr_2]`.
* We [encourage hyperlinks over footnotes for URLs](https://joss.readthedocs.io/en/latest/paper.html#links).
