### 2025-04-02 #7972 @KostasValeckas
Hi @warrickball , thank you.

I have made the following changes to the paper:

* Countries added to  affiliations.
* Citations grouped under same parenthesis
* Footnotes exchanged to hyperlinks for URL's
