+++ README.md
@@ -1,4 +1,11 @@
 **PyLongslit**
 
+
 All information is provided at:
 https://kostasvaleckas.github.io/PyLongslit/index.html
+
+[![status](https://joss.theoj.org/papers/81b63ae8448434cef0857dd835f1dcc1/status.svg)](https://joss.theoj.org/papers/81b63ae8448434cef0857dd835f1dcc1)
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15091603.svg)](https://doi.org/10.5281/zenodo.15091603)
+
+[![astro-ph](https://img.shields.io/badge/arXiv-astro--ph-blue)](https://arxiv.org/abs/2504.00973)
