### 2025-05-13 c58de9a
typo in objtrace docs fixed
