### 2025-05-18 99922cb
objcenter guess refined fixed to only look for a peak withing +/- 3 fwhm of user guess
