### 2025-05-29 4eb0a02
error adding bugfix in spatial flat model
