### 2025-06-15 ac17bca
started example config files
