### 2025-06-16 22daa73
more work on example configs
