### 2025-06-16 6edc6d7
docs and few fixes for closely spaced spectra extraction
