### 2025-06-16 c777c69
docs about updated instruments + path to make tests works with new database structure
