+++ _sources/index.rst.txt
@@ -3,6 +3,8 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
+.. _index:
+
 PyLongslit's documentation page
 ======================================
 
@@ -95,6 +97,7 @@ In this documentation, we provide guidance on installation and usage of the soft
 
    installation
    getting_started
+   tested_instruments
    special_use_cases
    develop
    uncertainties
