+++ _sources/tested_instruments.rst.txt
@@ -1,0 +1,78 @@
+.. _tested_instruments:
+
+Tested Instruments and Configurations
+=====================================
+
+As described in the :ref:`pipeline overview <index>`, the pipeline is designed 
+to be instrument-independent, as long as some primary assumptions are met 
+(see the :ref:`pipeline overview <index>` for details).
+
+However, the :ref:`configuration files <configuration_file>` for an instrument 
+setup can be viewed as an instrument-implementation, as for a fixed 
+instrument setup, most of the parameters in the :ref:`configuration file <configuration_file>`
+will be constant and very little will need to be changed between different datasets.
+Furthermore, resources like the products of :ref:`initial arc line identification <identify>` and 
+:ref:`extinction curves <sensfunction>` can be re-used. We therefore provide an 
+overview of already tested instrument setups with their resources and configuration files 
+in hope that these will be useful for users of the software.
+
+.. list-table::
+    :header-rows: 1
+    :widths: 25 25 25 25 25 25
+
+    * - Instrument
+      - Telescope
+      - Disperser
+      - Pixtable (Initial Arc Lines Guess)
+      - Extinction Curve
+      - Configuration File Example
+
+    * - ALFOSC
+      - Nordic Optical Telescope
+      - Grism #4
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/pixtables/alfosc_grating4_hene_pixtable.dat>`__ 
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/extinction_curves/lapalma.dat>`__ `Source <https://www.ing.iac.es/Astronomy/observing/manuals/html_manuals/wht_instr/pfip/node244.html>`__ 
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/example_configuration_files/ALFOSC/grism4.json>`__
+
+    * - ALFOSC
+      - Nordic Optical Telescope
+      - Grism #18
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/pixtables/alfosc_grating18_thar_pixtable.dat>`__
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/extinction_curves/lapalma.dat>`__ `Source <https://www.ing.iac.es/Astronomy/observing/manuals/html_manuals/wht_instr/pfip/node244.html>`__
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/example_configuration_files/ALFOSC/grism18.json>`__
+
+    * - ALFOSC
+      - Nordic Optical Telescope
+      - Grism #19
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/pixtables/alfosc_grating19_thar_pixtable.dat>`__
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/extinction_curves/lapalma.dat>`__ `Source <https://www.ing.iac.es/Astronomy/observing/manuals/html_manuals/wht_instr/pfip/node244.html>`__
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/example_configuration_files/ALFOSC/grism19.json>`__
+
+    * - ALFOSC
+      - Nordic Optical Telescope
+      - Grism #20
+      - `Link <https://github.com/KostasValeckas/PyLongslit_dev/blob/main/database/pixtables/alfosc_grating20_thar_pixtable.dat>`__
[…]
