+++ README.md
@@ -1,6 +1,5 @@
 **PyLongslit**
 
-
 All information is provided at:
 https://kostasvaleckas.github.io/PyLongslit/index.html
 
