+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = pylongslit
 author = Kostas Valeckas
-version = 1.0.6
+version = 1.0.7
 author_email = kostas.valeckas@nbi.ku.dk
 description = PyLongslit is a simple manual Python pipeline for processing astronomical long-slit spectra recorded with CCD detectors.
 long_description = file: README.md
