### 2025-07-22 e553015
implemneted a constant offset for better object tracing in A-B differenced images
