### 2025-08-21 e7f23f1
A few spelling mistakes in docs
