### 2025-08-21 eaca871
A few spelling mistakes in docs
