### 2025-09-02 f1f9416
Remove some unused dependencies and add [doc] and [test] requirments setting.
