+++ setup.cfg
@@ -21,81 +21,39 @@ classifiers=
 python_requires = >=3.10
 setup_requires = setuptools_scm
 install_requires =
+    astropy
+    astroscrappy
+    colorama
+    matplotlib
+    numpy
+    pandas
+    photutils
+    pyastronomy
+    PyQt5
+    pytest
+    Requests
+    scikit_learn
+    scipy
     setuptools
-    alabaster==1.0.0
-    astropy==6.1.7
-    astropy-iers-data==0.2025.2.24.0.34.4
-    astroscrappy==1.2.0
-    attrs==25.3.0
-    babel==2.17.0
-    bidict==0.23.1
-    cachetools==5.5.2
-    cattrs==24.1.2
-    certifi==2025.1.31
-    chardet==5.2.0
-    charset-normalizer==3.4.1
-    colorama==0.4.6
-    contourpy==1.3.1
-    cycler==0.12.1
-    distlib==0.3.9
-    docutils==0.21.2
-    esbonio==0.16.5
-    exceptiongroup==1.2.2
-    filelock==3.18.0
-    fonttools==4.56.0
-    idna==3.10
-    imageio==2.37.0
-    imagesize==1.4.1
-    iniconfig==2.0.0
-    Jinja2==3.1.6
-    joblib==1.4.2
-    kiwisolver==1.4.8
-    lazy_loader==0.4
-    lsprotocol==2023.0.1
-    MarkupSafe==3.0.2
-    matplotlib==3.8.0
-    networkx==3.4.2
-    numpy==1.26.4
-    photutils==2.0.2
-    pillow==11.1.0
-    platformdirs==4.3.6
-    pluggy==1.5.0
-    ply==3.11
-    PyAstronomy==0.22.0
-    pyerfa==2.0.1.5
-    pygls==1.3.1
-    Pygments==2.19.1
-    pyparsing==3.2.1
-    pyproject-api==1.9.0
-    PyQt5==5.15.10
-    pyspellchecker==0.8.2
-    pytest==8.3.5
-    pytest-order==1.3.0
-    python-dateutil==2.9.0.post0
-    PyYAML==6.0.2
-    quantities==0.16.1
-    requests==2.32.3
-    scikit-image==0.25.2
-    scikit-learn==1.6.1
-    scipy==1.15.2
-    six==1.17.0
-    snowballstemmer==2.2.0
-    Sphinx==8.1.3
-    sphinxcontrib-applehelp==2.0.0
-    sphinxcontrib-devhelp==2.0.0
-    sphinxcontrib-htmlhelp==2.1.0
-    sphinxcontrib-jsmath==1.0.1
-    sphinxcontrib-qthelp==2.0.0
-    sphinxcontrib-serializinghtml==2.0.0
-    threadpoolctl==3.5.0
-    tifffile==2025.2.18
-    tomli==2.2.1
-    tqdm==4.67.1
-    typing_extensions==4.12.2
-    urllib3==2.3.0
-    virtualenv==20.29.3
+    scikit-image
+    tabulate
+    tqdm
+        
 packages = find:
 include_package_data = True
+
+[options.extras_require]
+doc =
+    sphinx
+    sphinx-rtd-theme
+
+test =
+    pytest
+    pytest-order
+
+all =
+    %(doc)s
+    %(test)s
 
 [options.entry_points]
 console_scripts =
