### 2025-09-04 51f8e62
This is an add-on to PR (#26)
- Refined the proposed dependency cleanup
- Added docs to reflect changes
- Cleaned up some leftover-code
