+++ pylongslit/wavecalib.py
@@ -2368,7 +2368,6 @@ def main():
         description="Run the pylongslit wavecalibration procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
-    # Add more arguments as needed
 
     args = parser.parse_args()
 
