+++ pylongslit/extract_1d.py
@@ -570,7 +570,6 @@ def main():
         description="Run the pylongslit extract-1d procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
-    # Add more arguments as needed
 
     args = parser.parse_args()
 
