+++ setup.cfg
@@ -21,23 +21,20 @@ classifiers=
 python_requires = >=3.10
 setup_requires = setuptools_scm
 install_requires =
-    astropy
-    astroscrappy
-    colorama
-    matplotlib
-    numpy
-    pandas
-    photutils
-    pyastronomy
-    PyQt5
-    pytest
-    Requests
-    scikit_learn
-    scipy
     setuptools
-    scikit-image
-    tabulate
-    tqdm
+    astropy >= 6.1.7
+    astroscrappy >= 1.2.0
+    colorama >= 0.4.6
+    matplotlib >= 3.8.0
+    numpy >= 1.26.4, < 2.0.0
+    photutils >= 2.0.2
+    PyAstronomy >= 0.22.0
+    PyQt5 >= 5.15.10
+    scikit_learn >= 1.6.1
+    scipy >= 1.15.2
+    scikit-image >= 0.25.2
+    tqdm >= 4.67.1
+    PyAstronomy >= 0.22.0
         
 packages = find:
 include_package_data = True
@@ -46,10 +43,13 @@ include_package_data = True
 doc =
     sphinx
     sphinx-rtd-theme
+    sphinx-autodoc-typehints
+    myst-parser
 
 test =
     pytest
     pytest-order
+    requests
 
 all =
     %(doc)s
