### 2025-09-04 67de26b
This is an add-on to PR (#26)
- Refined the proposed dependency cleanup
- Added docs to reflect changes
- Cleaned up some leftover-code
