+++ pylongslit/reduce.py
@@ -418,7 +418,6 @@ def main():
         description="Run the pylongslit cosmic-ray removal procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
-    # Add more arguments as needed
 
     args = parser.parse_args()
 
