+++ pylongslit/crremoval.py
@@ -205,6 +205,7 @@ def main():
         description="Run the pylongslit cosmic-ray removal procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    # Add more arguments as needed
 
     args = parser.parse_args()
 
