+++ pylongslit/identify.py
@@ -24,26 +24,11 @@ from matplotlib.backends.backend_qt5agg import (
     NavigationToolbar2QT as NavigationToolbar,
 )
 from matplotlib.figure import Figure
+import matplotlib.pyplot as plt
 
-from PyQt5.QtCore import Qt, QRegExp
-from PyQt5.QtGui import QColor, QFont, QIntValidator, QRegExpValidator
-from PyQt5.QtWidgets import (
-    QAction,
-    QApplication,
-    QFileDialog,
-    QHBoxLayout,
-    QLabel,
-    QLineEdit,
-    QMainWindow,
-    QMessageBox,
-    QPushButton,
-    QSizePolicy,
-    QTableWidget,
-    QTableWidgetItem,
-    QToolBar,
-    QVBoxLayout,
-    QWidget
-)
+from PyQt5.QtCore import *
+from PyQt5.QtGui import *
+from PyQt5.QtWidgets import *
 
 from scipy.optimize import curve_fit
 from numpy.polynomial import Chebyshev
