+++ pylongslit/obj_trace_clone.py
@@ -331,6 +331,7 @@ def main():
         description="Run the pylongslit cloned object trace procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    # Add more arguments as needed
 
     args = parser.parse_args()
 
