+++ pylongslit/sensitivity_function.py
@@ -895,6 +895,7 @@ def main():
         description="Run the pylongslit sensitivity function procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    # Add more arguments as needed
 
     args = parser.parse_args()
 
