+++ pylongslit/obj_trace.py
@@ -1174,6 +1174,7 @@ def main():
         description="Run the pylongslit object-tracing procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    # Add more arguments as needed
 
     args = parser.parse_args()
 
