+++ pylongslit/subtract_background.py
@@ -193,6 +193,7 @@ def main():
         description="Run the pylongslit background-subtraction procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    # Add more arguments as needed
 
     args = parser.parse_args()
 
