### 2025-09-04 f022412
minor maintenance updates: gitignore, test workflows, requirments
