+++ requirements.txt
@@ -1,17 +1,14 @@
-astropy
-astroscrappy
-colorama
-matplotlib
-numpy
-pandas
-photutils
-pyastronomy
-PyQt5
-pytest
-Requests
-scikit_learn
-scipy
 setuptools
-scikit-image
-tabulate
-tqdm
+astropy >= 6.1.7
+astroscrappy >= 1.2.0
+colorama >= 0.4.6
+matplotlib >= 3.8.0
+numpy >= 1.26.4, < 2.0.0
+photutils >= 2.0.2
+PyAstronomy >= 0.22.0
+PyQt5 >= 5.15.10
+scikit_learn >= 1.6.1
+scipy >= 1.15.2
+scikit-image >= 0.25.2
+tqdm >= 4.67.1
+PyAstronomy >= 0.22.0
