### 2025-09-08 #7972 @KostasValeckas
Hi @warrickball . Thanks a lot for your answer, and no worries about the waiting time.

I will add some minor updates later this week. I also expect some research to soon be published where this software was used - I will add this to the article if it gets published before the review starts.

You do not need to pause the review before I have updated the paper. These will be minor changes, I just wanted to check whether the paper is "locked" after the submission, or if edits were allowed.

Thanks again,
Kostas
