### 2025-10-04 72d3530
prototype for combinning reduce 2d spectra implemented
