+++ setup.cfg
@@ -77,3 +77,4 @@ console_scripts =
     pylongslit_sensitivity_function = pylongslit.sensitivity_function:main
     pylongslit_flux = pylongslit.flux_calibrate:main
     pylongslit_combine_spec = pylongslit.combine:main
+    pylongslit_combine_2dspec = pylongslit.combine_2d:main
