### 2025-10-15 #7972 @warrickball
Thanks @kbwestfall! I'll tag you shortly and start the review.

You raise a good point that @KostasValeckas should address but I'm happy for this to be part of the review itself. In particular, our docs on the software paper in the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) say

> **State of the field**: Do the authors describe how this software compares to other commonly-used packages?
