### 2025-10-15 #9264 @editorialbot
[REVIEW]: PyLongslit: a simple manual Python pipeline for processing of astronomical long-slit spectra recorded with CCD detectors
**Submitting author:** @KostasValeckas (<a href="http://orcid.org/0009-0007-7275-0619">Kostas Valeckas</a>)
**Repository:** https://github.com/KostasValeckas/PyLongslit
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.5
