### 2025-11-06 d4bd33c
refactored docs on tutorial data
