### 2025-11-06 d650f49
few spelling errors fixed in docs
