### 2025-11-27 5669648
docs update - hyperlinked footer links
