### 2025-11-27 98f7e6d
docs fixed small leftover text
