### 2025-11-28 8b91e83
fixed matplotlib depency to avoid error in identify
