+++ requirements.txt
@@ -2,7 +2,7 @@ setuptools
 astropy >= 6.1.7
 astroscrappy >= 1.2.0
 colorama >= 0.4.6
-matplotlib >= 3.8.0
+matplotlib == 3.8.0
 numpy >= 1.26.4, < 2.0.0
 photutils >= 2.0.2
 PyAstronomy >= 0.22.0
