### 2025-11-28 85ed716
minor utils syntax change and fixed broken links in docs
