### 2025-11-29 5609c4f
full docs proof-read
