+++ docs/contact.rst
@@ -3,10 +3,10 @@
 Contact
 =======
 
-Any bug-reports, questions, or suggestions can be 
+Any bug reports, questions, or suggestions can be 
 directed through our 
 `GitHub repository <https://github.com/KostasValeckas/PyLongslit/issues>`_
-or by email to :
+or by email to:
 
 kostas.valeckas@nbi.ku.dk
 
