+++ docs/index.rst
@@ -19,7 +19,7 @@ PyLongslit's documentation page
 
 
 PyLongslit is a simple manual Python pipeline for processing of astronomical long-slit spectra 
-recorded with CCD detectors. At the current version, the software can produce following products:
+recorded with CCD detectors. At the current version, the software can produce the following products:
 
 #. A calibrated 2D spectrum in counts/wavelength.
 #. A 1D spectrum extracted from the 2D spectrum in counts/wavelength (for point-like objects).
@@ -35,19 +35,19 @@ is **designed to work with data that has the following characteristics**:
 #. The data is in counts, and has not been reduced in any way.
 #. Bias, flat-field and arc-lamp calibration frames are available. For the arc-lamp, accompanying
    data is available with a list of lines and their wavelengths, with the lines identified in the arc-lamp spectrum.
-   *(Wavelength-calibrating using sky-lines should be fine aswell, but it is not yet a tested approach).*
+   *(Wavelength-calibrating using sky-lines should be fine as well, but it is not yet a tested approach).*
 
 If these conditions are met, the software should be able to reduce the data to 2D and 1D spectra
 in counts/wavelength. For further processing, you will also need to acquire a standard star observation
 with a reference spectrum, and an extinction curve for your observatory.
 
 If your data does not meet these conditions, you can still try to use the software, but
-please know that the software is not designed/tested for data does not meet the above conditions.
+please know that the software is not designed/tested for data that does not meet the above conditions.
 
 **Principles behind the software:**
 
 #. We strive for **simplicity**. *The main motivation for the software is to create a tool that is
-   easy to understand and use, specially for novice astronomers and students. 
+   easy to understand and use, especially for novice astronomers and students. 
    There are many high-fidelity software solutions that are more sophisticated and 
    precise than PyLongslit, and we do not aim to out-perform them. Our goal is to
    create an entry-level pipeline that is approachable.*
@@ -59,11 +59,11 @@ please know that the software is not designed/tested for data does not meet the 
    processing, and do not compromise on this principle.*
 #. We strive for **transparency**. *We look at documentation as an active part of the code.
    No line of code is deemed complete if it is not documented appropriately.
-   We specially aim to disclose the limitations of the software, and to provide
-   guidance on how to interpret the results - specially through carefully designed quality-assesment plots.*
+   We especially aim to disclose the limitations of the software, and to provide
+   guidance on how to interpret the results - especially through carefully designed quality-assessment plots.*
 #. We strive for **robustness**. *We pay special attention to error handling and to
    providing guidance on how to troubleshoot common issues. We commit to providing 
-   support to users who are struggling with the software to our best ability.*
+   support to users who are struggling with the software to the best of our ability.*
 
 
 **Citing the software**
