### 2025-12-04 #9264 @KostasValeckas
Hi @kbwestfall and @Gabriel-p .

Thanks for your input.

@kbwestfall I am working on the issues now and will let you know as soon they are resolved.

@Gabriel-p can you please elaborate on your comment? Currently both JOSS and arXiv version of the PypeIt paper are referenced in PDF - as this is what the PypeIt team prefers according to their [docs](https://pypeit.readthedocs.io/en/stable/#citation) - do you mean that I should remove the arXiv DOI - or something else? Thank you.
