### 2025-12-05 9bc10e9
refs update for paper
