### 2025-12-05 55355c4
refs update for paper
