+++ paper/paper.bib
@@ -82,25 +82,6 @@ abstract = {An optimal spectrum extraction procedure is described, and examples 
   publisher={Cambridge University Press}
 }
 
-
-@ARTICLE{pypeit:joss_arXiv,
-       author = {{Prochaska}, J. Xavier and {Hennawi}, Joseph F. and {Westfall}, Kyle B. and
-         {Cooke}, Ryan J. and {Wang}, Feige and {Hsyu}, Tiffany and
-         {Davies}, Frederick B. and {Farina}, Emanuele Paolo},
-        title = "{PypeIt: The Python Spectroscopic Data Reduction Pipeline}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2020,
-        month = may,
-          eid = {arXiv:2005.06505},
-        pages = {arXiv:2005.06505},
-archivePrefix = {arXiv},
-       eprint = {2005.06505},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200506505P},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 @article{pypeit:joss_pub,
     doi = {10.21105/joss.02308},
     url = {https://doi.org/10.21105/joss.02308},
