### 2025-12-05 e32eed0
small paper update
