### 2025-12-05 06ad671
cleaned up  setup files
