### 2025-12-06 bd75a65
adressed all comments in issue #36
