### 2025-12-07 a894de1
adressed all comments in issue #37
