+++ pylongslit/wavecalib.py
@@ -268,7 +268,7 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
         Master arc image.
 
     center_row : int
-        Spacial index from where to start the fitting. User can shift it from the center, calling it "center-row" is just a convention.
+        spatial index from where to start the fitting. User can shift it from the center, calling it "center-row" is just a convention.
 
     fitter : `~astropy.modeling.fitting.LevMarLSQFitter`
         Fitter object.
@@ -284,8 +284,8 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
     all_centers : array
         Centers of the lines.
 
-    used_spacial : array
-        Spacial coordinates used for the fit of the line.
+    used_spatial : array
+        spatial coordinates used for the fit of the line.
 
     keep_mask : array
         Mask for keeping the good fits.
@@ -300,7 +300,7 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
     # further developtment or debugging.
     all_params = {}
 
-    # This amount of pixels is used +/- around the spacial pixel
+    # This amount of pixels is used +/- around the spatial pixel
     # in order to create a mean count array for the line (reduces noise).
     pixel_cut_extension = wavecalib_params["pixel_cut_extension"]
 
@@ -334,7 +334,7 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
     # new variable to deal with python indexing - the end pixel is not included
     cut_width = pixel_cut_extension + 1
 
-    # for book-keeping what index corresponds to what spacial pixel
+    # for book-keeping what index corresponds to what spatial pixel
     index_dict = {}
     num_it = 0
     for i in chain(
@@ -361,10 +361,10 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
     # the fitter checks that the next fit is not too far from the previous one
     last_good_center = g_model.mean_0.value
 
-    # container for what spacial pixels the fits were successful for
+    # container for what spatial pixels the fits were successful for
     used_spacial = []
 
-    # This is the driver loop - we loop over the spacial pixels from the center row
+    # This is the driver loop - we loop over the spatial pixels from the center row
     # and go in both directions until we reach the edge of the detector.
 
     for i in chain(
@@ -407,7 +407,7 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
                 print(start_pixel, end_pixel)
             continue
 
-        # extract the subimage around the line, and average over the spacial pixels
+        # extract the subimage around the line, and average over the spatial pixels
         center_row_spec = np.mean(
             master_arc.data[
                 i - pixel_cut_extension : i + pixel_cut_extension + 1,
@@ -492,7 +492,7 @@ def trace_line_tilt(master_arc, center_row, fitter, g_model, FWHM_guess):
                 # for several bad fits, might be that the center guess is off
                 # if enough good fits are present. Try to fit for the line tilt
                 # and use the tilt to estimate the center of the line at the current
-                # spacial row
+                # spatial row
 
                 # at least 10 good fits are needed to try this - hardcoded value for now
                 if np.sum(keep_mask) > 10:
[…]
