### 2025-12-07 47fae6d
small docs typos and version increment
