+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = pylongslit
 author = Kostas Valeckas
-version = 1.1.3
+version = 1.1.4
 author_email = kostas.valeckas@nbi.ku.dk
 description = PyLongslit is a simple manual Python pipeline for processing astronomical long-slit spectra recorded with CCD detectors.
 long_description = file: README.md
@@ -25,7 +25,7 @@ install_requires =
     astropy >= 6.1.7
     astroscrappy >= 1.2.0
     colorama >= 0.4.6
-    matplotlib >= 3.10.7 
+    matplotlib >= 3.10.7
     numpy >= 1.26.4, < 2.0.0
     photutils >= 2.0.2
     PyAstronomy >= 0.22.0
