### 2025-12-07 60e325e
implemented version -v option to all console commands
