+++ pylongslit/reduce.py
@@ -414,10 +414,16 @@ def reduce_all():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(
         description="Run the pylongslit cosmic-ray removal procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
