+++ pylongslit/combine.py
@@ -226,10 +226,16 @@ def run_combine_spec():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(
         description="Run the pylongslit combine fluxed spectrum procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
