+++ pylongslit/flux_calibrate.py
@@ -259,10 +259,16 @@ def run_flux_calib():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(
         description="Run the pylongslit flux-calibration procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
