+++ pylongslit/mkspecbias.py
@@ -106,8 +106,14 @@ def run_bias():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(description="Run the pylongslit bias procedure.")
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
