+++ pylongslit/combine_arcs.py
@@ -115,11 +115,17 @@ def combine_arcs():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(
         description="Run the pylongslit combine-arc procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
     # Add more arguments as needed
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
