+++ pylongslit/check_config.py
@@ -384,10 +384,18 @@ def run_config_checks():
 
 
 def main():
+
+    from pylongslit.version import get_version
+
     parser = argparse.ArgumentParser(
         description="Run the pylongslit config-file checker."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+        "-v", "--version", 
+        action="version", 
+        version=f"PyLongslit {get_version()}"
+    )
 
     args = parser.parse_args()
 
