+++ pylongslit/mkspecflat.py
@@ -967,10 +967,16 @@ def run_flats():
 
 
 def main():
+    from pylongslit.version import get_version
     parser = argparse.ArgumentParser(
         description="Run the pylongslit flatfield procedure."
     )
     parser.add_argument("config", type=str, help="Configuration file path")
+    parser.add_argument(
+    "-v", "--version", 
+    action="version", 
+    version=f"PyLongslit {get_version()}"
+)
 
     args = parser.parse_args()
 
