### 2025-12-07 c5dc10f
Update pylongslit/check_config.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
