### 2025-12-07 dfc42ea
Update pylongslit/version.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
