+++ pylongslit/version.py
@@ -1,12 +1,8 @@
+from importlib.metadata import version, PackageNotFoundError
+
 def get_version():
-    """Get the version from setuptools metadata or fallback."""
+    """Get the version from setuptools metadata."""
     try:
-        from importlib.metadata import version
         return version("pylongslit")
-    except ImportError:
-        # Fallback for Python < 3.8
-        try:
-            import pkg_resources
-            return pkg_resources.get_distribution("pylongslit").version
-        except:
-            return "unknown"
+    except PackageNotFoundError:
+        return "unknown"
