### 2025-12-07 dfeac83
paper update and version increment
