### 2025-12-07 #9264 @KostasValeckas
Hi @Gabriel-p and @kbwestfall .

I have now taken action on all of your comments both from this thread and the issues in the repository. I have left the issues open for further discussions if needed.

As I have found all of the comments very useful I have pulled them to the main branch and deployed them to production straight away. The newest version of the software is 1.1.5.  For bigger changes, I have linked the PR's to the Issues so it is hopefully easier for you to see what exactly has been changed.

I still need to address the issue @Gabriel-p has submitted about loading the pixtables from the identify routine - but I need some further information, as I am not successful in recreating the error on my own machine (Linux Ubuntu) or in a MacOS VM. I have left a comment on the issue requesting some further info.

Regarding adding a footnote to pypeit repo - I have added it as hypertext instead as I got a comment in the pre-review to not add URL's as footnotes.

I will generate the updated paper PDF.
