### 2025-12-07 fd36c6c
minor typos in paper
