### 2025-12-09 #9264 @KostasValeckas
Thanks @warrickball  !

I have made a minor affiliation update to the paper (added my affiliation to the Nordic Optical Telescope).

The version number for release is the current v1.1.5 .

I have another version of the software on Zenodo, so I have just uploaded v.1.1.5 so it gets its' own DOI: [10.5281/zenodo.15306323](https://doi.org/10.5281/zenodo.15306323) . Please let me know if this is an allowed approach.

Double-checked authors, ORCIDs, affiliations etc.

Please let me know if further action is needed.
