### 2025-12-14 607e748
Initial editorial changes for publication in JOSS
