+++ paper/paper.md
@@ -19,12 +19,10 @@ authors:
     orcid: 0000-0002-9389-7413
     affiliation: 2
 
-
-
 affiliations:
  - name: Niels Bohr Institute, Copenhagen University, Denmark
    index: 1
-  
+
 # - name: Nordic Optical Telescope
 #   index: 2
 
@@ -36,7 +34,6 @@ affiliations:
 
  - name: Nordic Optical Telescope, Spain
    index: 4
-
 
 date: 27 March 2025
 bibliography: paper.bib
@@ -52,9 +49,9 @@ bibliography: paper.bib
 We present a new Python package for processing data from astronomical 
 long-slit spectroscopy observations recorded with CCD detectors.
 
-The software is designed to aim for **simplicity**, **manual execution**, **transparency** and **robustness**. The goal for the software is to provide a manual and simple counterpart to the 
-well-established semi-automated and automated pipelines. The intended use-cases are **teaching** and **cases where 
-automated pipelines fail**. For further elaboration,
+The software is designed to aim for simplicity, manual execution, transparency and robustness. The goal for the software is to provide a manual and simple counterpart to the 
+well-established semi-automated and automated pipelines. The intended use cases are teaching and cases where 
+automated pipelines fail. For further elaboration,
 please see the [Statement of need](#statement-of-need). 
 
 From raw data, the
@@ -71,17 +68,17 @@ detector calibrations [@handbook ; @Howell_2006], cosmic-ray subtraction [@cr_1 
 
 # Statement of need
 
-A natural approach when developing data-processing pipelines is to seek for precision and automation. The trade-off for this 
+A natural approach when developing data processing pipelines is to seek for precision and automation. The trade off for this 
 is code complexity and "black-box" solutions, where the process of the pipeline is often masked, and 
-the quality-assessment output is made under the assumption that the user knows how to interpret it. 
-In research, this is a reasonable trade-off, as a certain level of user-skill and experience can be assumed. However, 
+the quality assessment output is made under the assumption that the user knows how to interpret it. 
+In research, this is a reasonable trade off, as a certain level of user skill and experience can be assumed. However, 
 in a teaching paradigm, simplicity and transparency are often more favorable, even when this means loss of 
 precision and automation. The PyLongslit pipeline is designed to rely on simple code and manual execution,
-supported by a large array of quality-assessment plots and extensive documentation. The algorithms are designed to produce research-quality results, yet while prioritizing simplicity over high precision. The reason for this is 
+supported by a large array of quality assessment plots and extensive documentation. The algorithms are designed to produce research-quality results, yet while prioritizing simplicity over high precision. The reason for this is 
 to create a robust and transparent pipeline, where every step of the execution is visualized and explained. We see this as being especially valuable in teaching scenarios and for users who are new to spectroscopic data processing. Furthermore, we hope that 
-the simple coding style will invite users of all skill-levels to contribute to the code.
+the simple coding style will invite users of all skill levels to contribute to the code.
 
-An early beta-version of the software was user-tested during the [Nordic Optical Telescope](https://www.not.iac.es/) [IDA summer-course 
+An early beta version of the software was user tested during the [Nordic Optical Telescope](https://www.not.iac.es/) [IDA summer-course 
 2024](https://phys.au.dk/ida/events/not-summer-school-2024), where all student groups were able to follow the documentation and successfully process data 
 without any significant assistance. 
 
@@ -89,22 +86,22 @@ During the development of software it became apparent that the manual nature of 
[…]
