+++ paper/paper.bib
@@ -27,7 +27,7 @@ volume = {98},
 number = {604},
 pages = {609},
 author = {Horne, K.},
-title = {AN OPTIMAL EXTRACTION ALGORITHM FOR CCD SPECTROSCOPY.},
+title = {An optimal extraction algorithm for CCD spectroscopy.},
 journal = {Publications of the Astronomical Society of the Pacific},
 abstract = {An optimal spectrum extraction procedure is described, and examples of its performance with CCD data are presented. The algorithm delivers the maximum possible signal-to-noise ratio while preserving spectrophotometric accuracy. The effects of moderate geometric distortion and of cosmic-ray hits on the spectrum are automatically accounted for. In tests with background-noise limited CCD spectra, optimal extraction offers a 70-percent gain in effective exposure time in comparison with conventional extraction procedures.}
 }
@@ -45,7 +45,7 @@ abstract = {An optimal spectrum extraction procedure is described, and examples 
                   Brigitta Sipocz and
                   Thomas Robitaille and
                   Christoph Deil},
-  title        = {astropy/astroscrappy: v1.0.5 Zenodo Release},
+  title        = {astropy/astroscrappy: v1.0.5},
   month        = nov,
   year         = 2018,
   publisher    = {Zenodo},
@@ -55,7 +55,7 @@ abstract = {An optimal spectrum extraction procedure is described, and examples 
 }
 
 @article{cr_2,
-   title={Cosmic‐Ray Rejection by Laplacian Edge Detection},
+   title={Cosmic‐Ray Rejection by {Laplacian} Edge Detection},
    volume={113},
    ISSN={1538-3873},
    url={http://dx.doi.org/10.1086/323894},
