### 2025-12-16 d4daa97
minor paper bib. update
