+++ paper/paper.bib
@@ -91,7 +91,7 @@ abstract = {An optimal spectrum extraction procedure is described, and examples 
     number = {56},
     pages = {2308},
     author = {J. Xavier Prochaska and Joseph F. Hennawi and Kyle B. Westfall and Ryan J. Cooke and Feige Wang and Tiffany Hsyu and Frederick B. Davies and Emanuele Paolo Farina and Debora Pelliccia},
-    title = {PypeIt: The Python Spectroscopic Data Reduction Pipeline},
+    title = {PypeIt: The {Python} Spectroscopic Data Reduction Pipeline},
     journal = {Journal of Open Source Software}
 }
 
