### 2025-12-16 #9264 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
