### 2025-12-16 #9264 @warrickball
Thanks once again to @Gabriel-p & @kbwestfall for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
